refactor: move minting functionality to minting extension folder
